Remove old folder and move code
