+++ vc_issuance_contract/src/base32.rs
