+++ vc_issuance_contract/src/contract.rs
@@ -1,8 +1,10 @@
-use crate::error::ContractError;
 use crate::storage;
-
+use crate::vault_contract;
 use crate::vc_issuance_trait::VCIssuanceTrait;
-use soroban_sdk::{contract, contractimpl, contractmeta, panic_with_error, Address, Env, Map, Vec};
+use crate::{error::ContractError, verifiable_credential};
+use soroban_sdk::{
+    contract, contractimpl, contractmeta, panic_with_error, Address, Env, Map, String, Vec,
+};
 
 const LEDGERS_THRESHOLD: u32 = 1;
 const LEDGERS_TO_EXTEND: u32 = 535_000;
@@ -38,4 +40,29 @@ impl VCIssuanceTrait for VCIssuanceContract {
             .instance()
             .bump(LEDGERS_THRESHOLD, LEDGERS_TO_EXTEND);
     }
+    fn issue(
+        e: Env,
+        admin: Address,
+        vc_data: String,
+        recipient_did: String,
+        storage_address: Address,
+    ) -> String {
+        validate_admin(&e, &admin);
+
+        let vc_id = verifiable_credential::generate_id(&e);
+
+        // let client = vault_contract::Client::new(&e, &storage_address);
+        // client.store_vc(&vc_id, &vc_data, &recipient_did, &admin, &storage_address);
+        verifiable_credential::add_vc(&e, &vc_id);
+
+        vc_id
+    }
+}
+
+fn validate_admin(e: &Env, admin: &Address) {
+    let contract_admin = storage::read_admin(e);
+    if contract_admin != admin.clone() {
+        panic_with_error!(e, ContractError::NotAuthorized)
+    }
+    admin.require_auth();
 }
