+++ vc_issuance_contract/src/storage.rs
@@ -15,6 +15,11 @@ pub fn has_admin(e: &Env) -> bool {
     e.storage().instance().has(&key)
 }
 
+pub fn read_admin(e: &Env) -> Address {
+    let key = DataKey::Admin;
+    e.storage().instance().get(&key).unwrap()
+}
+
 pub fn write_admin(e: &Env, id: &Address) {
     let key = DataKey::Admin;
     e.storage().instance().set(&key, id);
@@ -28,6 +33,11 @@ pub fn write_amount(e: &Env, amount: &u32) {
 pub fn write_vcs(e: &Env, vc: &Vec<String>) {
     let key = DataKey::VerifiableCredentials;
     e.storage().instance().set(&key, vc)
+}
+
+pub fn read_vcs(e: &Env) -> Vec<String> {
+    let key = DataKey::VerifiableCredentials;
+    e.storage().instance().get(&key).unwrap()
 }
 
 pub fn write_vcs_revocations(e: &Env, revocations: &Map<String, Revocation>) {
