+++ vc_issuance_contract/src/test/contract.rs
@@ -1,4 +1,6 @@
 use crate::test::setup::VCIssuanceContractTest;
+use crate::vault_contract;
+use soroban_sdk::{testutils::Address as _, vec, Address, String};
 
 #[test]
 fn test_initialize_with_amount() {
@@ -50,4 +52,28 @@ fn test_initialize_an_already_initialized_contract() {
 
     contract.initialize(&admin, &amount);
     contract.initialize(&admin, &amount);
+}
+
+#[test]
+fn test_issue() {
+    let VCIssuanceContractTest {
+        env,
+        admin,
+        amount,
+        contract,
+    } = VCIssuanceContractTest::setup();
+
+    let vc_data = String::from_slice(&env, "vc_data");
+    let recipient_did = String::from_slice(&env, "recipient_did");
+    let vault_admin = Address::random(&env);
+
+    let vault_contract_id = env.register_contract_wasm(None, vault_contract::WASM);
+    let vault_client = vault_contract::Client::new(&env, &vault_contract_id);
+    let dids = vec![&env, recipient_did.clone()];
+
+    vault_client.initialize(&vault_admin, &dids);
+    vault_client.authorize_issuer(&vault_admin, &admin, &recipient_did);
+
+    contract.initialize(&admin, &amount);
+    contract.issue(&admin, &vc_data, &recipient_did, &vault_contract_id);
 }
