+++ vc_issuance_contract/src/vc_issuance_trait.rs
@@ -1,6 +1,14 @@
-use soroban_sdk::{Address, Env};
+use soroban_sdk::{Address, Env, String};
 
 pub trait VCIssuanceTrait {
     /// Initializes the Verifiable Credentials Issuance Contract by setting the admin and an optional amount.
     fn initialize(e: Env, admin: Address, amount: Option<u32>);
+
+    fn issue(
+        e: Env,
+        admin: Address,
+        recipient_did: String,
+        vc_data: String,
+        storage_address: Address,
+    ) -> String;
 }
