+++ vc_issuance_contract/src/verifiable_credential.rs
