--- verifiable_credentials_contract/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "verifiable_credentials_contract"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-soroban-sdk = { workspace = true }
-
-[dev_dependencies]
-soroban-sdk = { workspace = true, features = ["testutils"] }
