--- verifiable_credentials_contract/src/error.rs
@@ -1,10 +1,0 @@
-use soroban_sdk::contracterror;
-
-#[contracterror]
-#[derive(Copy, Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
-#[repr(u32)]
-pub enum ContractError {
-    AlreadyInitialized = 1,
-    NotAuthorized = 2,
-    AmountLimitExceeded = 3,
-}
