--- verifiable_credentials_contract/src/revocation.rs
@@ -1,8 +1,0 @@
-use soroban_sdk::{contracttype, String};
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct Revocation {
-    pub vc_id: String,
-    pub date: String,
-}
