--- verifiable_credentials_contract/src/test.rs
@@ -1,2 +1,0 @@
-mod contract;
-mod setup;
